Seperating out infraoverview.api.js to different files based on node feature
